- Fixed the mess with label replacements in override transpiler (don't code on drugs, kids !) - Refactored the FI & VesselPre patches to make them MFI-compliant. They should no longer interfere with mods using MFI such as FAR.
