+++ GameData/KSPCommunityFixes/Settings.cfg
@@ -420,9 +420,13 @@ KSP_COMMUNITY_FIXES
   // state synchronization and caching solar panels scaled space raycasts results.
   OptimizedModuleRaycasts = true
 
-  // General micro-optimization of FlightIntegrator, VesselPrecalculate and FloatingOrigin, significantely 
-  // increase framerate in large part count situations.
-  FlightCoreSystemsPerf = true
+  // General micro-optimization of FlightIntegrator and VesselPrecalculate. This has a significant impact in
+  // large part count situations, and especially in atmospheric situations.
+  FlightIntegratorPerf = true
+
+  // General micro-optimization of floating origin shifts. Main benefit is in large particle count situations
+  // but this helps a bit in other cases as well.
+  FloatingOriginPerf = true
 
   // Faster lookup of other docking nodes
   ModuleDockingNodeFindOtherNodesFaster = true
@@ -434,7 +438,7 @@ KSP_COMMUNITY_FIXES
   // volume normalizer, strut position tracking...
   PartSystemsFastUpdate = true
 
-  // Various small patches on hot loop methods
+  // Various small performance patches
   MinorPerfTweaks = true
 
   // ##########################
