+++ KSPCommunityFixes/Performance/MinorPerfTweaks.cs
@@ -1,5 +1,4 @@
-﻿using HarmonyLib;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using UnityEngine;
